change return type of vmMain to intptr_t

+++ src/menu/ui_main.c
@@ -36,7 +36,7 @@ void _UI_Refresh( int realtime );
 int MVAPI_Init(int apilevel);
 
 qboolean _UI_IsFullscreen( void );
-LIBEXPORT int QDECL vmMain(int command, int arg0, int arg1, int arg2, int arg3, int arg4, int arg5, int arg6, int arg7, int arg8, int arg9, int arg10, int arg11) {
+LIBEXPORT intptr_t QDECL vmMain(int command, int arg0, int arg1, int arg2, int arg3, int arg4, int arg5, int arg6, int arg7, int arg8, int arg9, int arg10, int arg11) {
   switch ( command ) {
 	  case UI_GETAPIVERSION:
 		  return UI_API_16_VERSION;
